[ADVAPP-388] [Cleanup]: Introduce the ability to order prospect status, and default to the value in the create prospect experience to the highest ordered option
