+++ app-modules/prospect/src/Filament/Resources/ProspectResource/Pages/CreateProspect.php
@@ -38,7 +38,6 @@ namespace AdvisingApp\Prospect\Filament\Resources\ProspectResource\Pages;
 
 use App\Models\User;
 use Filament\Forms\Form;
-use Laravel\Pennant\Feature;
 use App\Models\Scopes\HasLicense;
 use Filament\Forms\Components\Radio;
 use Filament\Forms\Components\Select;
@@ -48,7 +47,6 @@ use AdvisingApp\Prospect\Models\Prospect;
 use Filament\Forms\Components\DatePicker;
 use Illuminate\Database\Eloquent\Builder;
 use Filament\Resources\Pages\CreateRecord;
-use App\Features\ProspectStatusSortFeature;
 use AdvisingApp\Prospect\Models\ProspectSource;
 use AdvisingApp\Prospect\Models\ProspectStatus;
 use AdvisingApp\Prospect\Filament\Resources\ProspectResource;
@@ -64,9 +62,7 @@ class CreateProspect extends CreateRecord
                 Select::make('status_id')
                     ->label('Status')
                     ->required()
-                    ->relationship('status', 'name', Feature::active(
-                        ProspectStatusSortFeature::class
-                    ) ? fn (Builder $query) => $query->orderBy('sort') : null)
+                    ->relationship('status', 'name', fn (Builder $query) => $query->orderBy('sort'))
                     ->default(ProspectStatus::query()->orderBy('sort')->first()?->getKey())
                     ->exists(
                         table: (new ProspectStatus())->getTable(),
