--- database/operations/2024_02_11_132430_fill_prospect_status_sort_values.php
@@ -1,79 +1,0 @@
-<?php
-
-/*
-<COPYRIGHT>
-
-    Copyright © 2016-2024, Canyon GBS LLC. All rights reserved.
-
-    Advising App™ is licensed under the Elastic License 2.0. For more details,
-    see https://github.com/canyongbs/advisingapp/blob/main/LICENSE.
-
-    Notice:
-
-    - You may not provide the software to third parties as a hosted or managed
-      service, where the service provides users with access to any substantial set of
-      the features or functionality of the software.
-    - You may not move, change, disable, or circumvent the license key functionality
-      in the software, and you may not remove or obscure any functionality in the
-      software that is protected by the license key.
-    - You may not alter, remove, or obscure any licensing, copyright, or other notices
-      of the licensor in the software. Any use of the licensor’s trademarks is subject
-      to applicable law.
-    - Canyon GBS LLC respects the intellectual property rights of others and expects the
-      same in return. Canyon GBS™ and Advising App™ are registered trademarks of
-      Canyon GBS LLC, and we are committed to enforcing and protecting our trademarks
-      vigorously.
-    - The software solution, including services, infrastructure, and code, is offered as a
-      Software as a Service (SaaS) by Canyon GBS LLC.
-    - Use of this software implies agreement to the license terms and conditions as stated
-      in the Elastic License 2.0.
-
-    For more information or inquiries please visit our website at
-    https://www.canyongbs.com or contact us via email at legal@canyongbs.com.
-
-</COPYRIGHT>
-*/
-
-use Laravel\Pennant\Feature;
-use App\Features\ProspectStatusSortFeature;
-use AdvisingApp\DataMigration\OneTimeOperation;
-use AdvisingApp\Prospect\Models\ProspectStatus;
-use AdvisingApp\DataMigration\Enums\OperationType;
-
-return new class () extends OneTimeOperation {
-    /**
-     * The type to determine where it will be run. OperationType::Tenant or OperationType::Landlord.
-     */
-    protected OperationType $type = OperationType::Tenant;
-
-    /**
-     * Determine if the operation is being processed asynchronously.
-     */
-    protected bool $async = true;
-
-    /**
-     * The queue that the job will be dispatched to. Will default to defaults in config.
-     */
-    protected ?string $queue = null;
-
-    /**
-     * A tag name, that this operation can be filtered by.
-     */
-    protected ?string $tag = 'after-deployment';
-
-    /**
-     * Process the operation.
-     */
-    public function process(): void
-    {
-        $sort = 1;
-
-        ProspectStatus::query()->orderBy('sort')->each(function (ProspectStatus $status) use (&$sort) {
-            $status->update(['sort' => $sort]);
-
-            $sort++;
-        });
-
-        Feature::activate(ProspectStatusSortFeature::class);
-    }
-};
