+++ database/operations/2024_02_11_132431_purge_prospect_status_sort_feature_flag.php
@@ -34,12 +34,36 @@
 </COPYRIGHT>
 */
 
-namespace App\Features;
+use Laravel\Pennant\Feature;
+use AdvisingApp\DataMigration\OneTimeOperation;
+use AdvisingApp\DataMigration\Enums\OperationType;
 
-class ProspectStatusSortFeature
-{
-    public function resolve(mixed $scope): mixed
+return new class () extends OneTimeOperation {
+    /**
+     * The type to determine where it will be run. OperationType::Tenant or OperationType::Landlord.
+     */
+    protected OperationType $type = OperationType::Tenant;
+
+    /**
+     * Determine if the operation is being processed asynchronously.
+     */
+    protected bool $async = true;
+
+    /**
+     * The queue that the job will be dispatched to. Will default to defaults in config.
+     */
+    protected ?string $queue = null;
+
+    /**
+     * A tag name, that this operation can be filtered by.
+     */
+    protected ?string $tag = 'after-deployment';
+
+    /**
+     * Process the operation.
+     */
+    public function process(): void
     {
-        return false;
+        Feature::purge('App\\Features\\ProspectStatusSortFeature');
     }
-}
+};
